requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
entrypoints==0.3
flake8==3.8.4
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
tabula-py==2.2.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
